+++ src/components/App/App.js
@@ -1,12 +1,12 @@
 import { Navigate, Route, Routes } from 'react-router-dom';
+import Login from '../Login/Login';
+import Main from '../Main/Main';
+import Movies from '../Movies/Movies';
+import SavedMovies from '../Movies/SavedMovies/SavedMovies';
+import NotFound from '../NotFound/NotFound';
+import Profile from '../Profile/Profile';
+import Register from '../Register/Register';
 import './App.css';
-import Login from './components/Login/Login';
-import Main from './components/Main/Main';
-import Movies from './components/Movies/Movies';
-import SavedMovies from './components/Movies/SavedMovies/SavedMovies';
-import NotFound from './components/NotFound/NotFound';
-import Profile from './components/Profile/Profile';
-import Register from './components/Register/Register';
 
 function App() {
   return (
